Split temporal tests into separate classes by type
